+++ AUTHORS
@@ -1,4 +1,4 @@
-# This is the official list of fa-scrapper authors for copyright purposes.
+# This is the official list of fa-scraper authors for copyright purposes.
 
 Pablo Baeyens (mx-psi) <pbaeyens31+github@gmail.com>
 Pau Riera (paurieraf) <pau.riera.forteza@gmail.com>
