+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 
 <!--- The following link definition list is generated by PimpMyChangelog --->
 
-[#21]: https://github.com/mx-psi/fa-scrapper/issues/21
-[#42]: https://github.com/mx-psi/fa-scrapper/issues/42
-[#43]: https://github.com/mx-psi/fa-scrapper/issues/43
+[#21]: https://github.com/mx-psi/fa-scraper/issues/21
+[#42]: https://github.com/mx-psi/fa-scraper/issues/42
+[#43]: https://github.com/mx-psi/fa-scraper/issues/43
 [@darruko]: https://github.com/Darruko
