+++ Dockerfile
@@ -5,4 +5,4 @@ ENV PYTHONDONTWRITEBYTECODE 1
 ENV PYTHONUNBUFFERED 1
 
 # Install package
-RUN pip3 install fa-scrapper
+RUN pip3 install fa-scraper
