+++ fa_scraper/types.py
